+++ packages/dso-toolkit/src/components/helpcenter-panel/helpcenter-panel.models.ts
@@ -1,4 +1,0 @@
-export interface HelpcenterPanel {
-  label?: string;
-  url: string;
-}