+++ CHANGELOG.md
@@ -11,6 +11,9 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) a
 ### Changed
 * Verticaal Ritme: afstand boven en onder knoppenbalk naar 32px ([#3013](https://github.com/dso-toolkit/dso-toolkit/issues/3013))
 
+### Removed
+* **BREAKING** Helpcentrum Panel: Remove ([#3068](https://github.com/dso-toolkit/dso-toolkit/issues/3068))
+
 ### Task
 * Packages: Dependencies updates ([#2977](https://github.com/dso-toolkit/dso-toolkit/issues/2977))